Pass the field values through .strip() before translating them.
This fixes #17 but may require some modification of translations
which rely on the spaces being there.

For example, a translation which did a check like

    if attrs['foo'] == '     bar  ':

will now need to do it as

    if attrs['foo'] == 'bar':

Checks of this type were unreliable anyways as they relied
on the spaces which only happened with some file formats. One
backwards compatible way to do the check that works is

    if attrs['foo'].strip() == 'bar':

This has the additional cost of calling strip() twice