add a LICENSE file containing the MIT license text, as is necessary
As far as I can discern Paul's intent is that all of this code is licensed as
MIT, but that license requires a copy of the license text to be provided with
the source. Attribution copied from ogr2osm.py, and files that mention the
license adjusted to refer to the LICENSE file rather than an online copy.